requirements.txt
certifi==2023.5.7
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
docutils==0.20.1
fonttools==4.40.0
grafana-client==3.5.0
idna==3.4
importlib-resources==6.0.0
Kivy==2.2.1
Kivy-Garden==0.1.5
kivy-garden.matplotlib @ https://github.com/g-braeunlich/garden.matplotlib/archive/pkg.zip
kiwisolver==1.4.4
matplotlib==3.7.2
numpy==1.25.1
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.6
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
urllib3==2.0.3
zipp==3.16.0
